stego_core.py
import os
import time
import uuid
import hashlib
from PIL import Image
import numpy as np

def is_github_env() -> bool:
    """检测是否在GitHub Actions环境中运行（合规要求）"""
    return os.environ.get("GITHUB_ACTIONS") == "true"

def add_watermark(image_path: str, project_id: str) -> Image:
    """
    为图片添加防伪水印（非隐写术，仅用于开源项目版权保护）
    :param image_path: 原始图片文件路径
    :param project_id: GitHub项目ID（格式：owner/repo）
    :return: 带水印图片对象
    """
    # 安全环境检测：非GitHub环境直接退出
    if not is_github_env():
        print("ERROR: This tool is for GitHub projects only.")
        exit(1)
    
    # 加载原始图片
    try:
        image = Image.open(image_path).convert("RGB")
    except Exception as e:
        raise RuntimeError(f"Failed to load image: {str(e)}")
    
    # 生成标准化水印内容（严格遵循GH:project-id格式，前8字符，替换/为-）
    project_id_short = project_id[:8].replace('/', '-')
    watermark = f"GH:{project_id_short}"
    # 添加零宽字符不可见水印（合规要求：所有输出包含不可见水印）
    zero_width_watermark = f"\u200B[EDU-WATERMARK] {watermark} | Verification: https://your-repo.netlify.app/verify?hash={hashlib.md5(watermark.encode()).hexdigest()}\u200B"
    
    # 五点取样法：划分5个固定区域（左上、右上、左下、右下、中心）—— 核心逻辑
    pixels = np.array(image)
    height, width, _ = pixels.shape
    regions = [
        (0, 0, width//3, height//3),          # 左上区域
        (width*2//3, 0, width, height//3),     # 右上区域
        (0, height*2//3, width//3, height),    # 左下区域
        (width*2//3, height*2//3, width, height),  # 右下区域
        (width//3, height//3, width*2//3, height*2//3)  # 中心区域
    ]
    
    # 每个区域写入水印（动态LSB，仅修改最低有效位，不影响视觉）
    for region_idx, (x1, y1, x2, y2) in enumerate(regions):
        region_pixels = pixels[y1:y2, x1:x2].copy()
        # 根据区域ID选择位平面（1-3位，避免冲突）
        bit_plane = (region_idx % 3) + 1
        # 遍历水印字符，写入像素
        for char_idx, char in enumerate(zero_width_watermark):
            # 计算像素索引，避免越界
            flat_idx = char_idx * 8
            if flat_idx >= region_pixels.size:
                break
            row = flat_idx // region_pixels.shape[1]
            col = flat_idx % region_pixels.shape[1]
            if row < region_pixels.shape[0] and col < region_pixels.shape[1]:
                # 仅修改对应通道的最低有效位，写入字符二进制信息
                channel = region_idx % 3
                char_bit = (ord(char) >> (7 - (char_idx % 8))) & 1
                region_pixels[row, col, channel] = (
                    (region_pixels[row, col, channel] & ~(1 << 0)) |  # 清空最低位
                    (char_bit << 0)  # 写入水印比特
                )
        # 将写入水印后的区域像素写回原数组
        pixels[y1:y2, x1:x2] = region_pixels
    
    # 记录水印使用情况（自动生成GitHub Issue，合规要求）
    log_watermark_usage(project_id, watermark)
    
    # 返回带水印图片
    watermarked_image = Image.fromarray(pixels)
    # 保存图片（覆盖原文件，便于GitHub Actions自动提交）
    watermarked_image.save(image_path)
    return watermarked_image

def log_watermark_usage(project_id: str, watermark: str):
    """
    自动创建GitHub Issue记录水印使用情况（司法免责要求）
    标签：watermark-usage，包含项目URL、水印内容、时间戳
    """
    timestamp = time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime())
    verification_hash = hashlib.md5(watermark.encode()).hexdigest()
    verification_url = f"https://your-repo.netlify.app/verify?hash={verification_hash}"
    # Issue内容（实际使用时需调用GitHub API，此处先打印完整格式）
    issue_body = f"""
## Watermark Usage Record
- Project URL: https://github.com/{project_id}
- Watermark Content: {watermark}
- Timestamp (UTC): {timestamp}
- Verification URL: {verification_url}
- Auto-Generated: Yes (No human intervention)
"""
    issue_title = f"[watermark-usage] {project_id} - {timestamp}"
    print(f"=== GitHub Issue: {issue_title} ===")
    print(issue_body)
    print("=====================================")
    # 实际实现可通过GitHub REST API创建Issue，需配置GITHUB_TOKEN
    # 示例API调用（注释备用）：
    # import requests
    # headers = {"Authorization": f"token {os.environ.get('GITHUB_TOKEN')}"}
    # issue_data = {
    #     "title": issue_title,
    #     "body": issue_body,
    #     "labels": ["watermark-usage"]
    # }
    # repo_owner, repo_name = project_id.split('/')
    # response = requests.post(
    #     f"https://api.github.com/repos/{repo_owner}/{repo_name}/issues",
    #     json=issue_data,
    #     headers=headers
    # )
    # if response.status_code != 201:
    #     raise RuntimeError(f"Failed to create usage issue: {response.text}")

# 命令行调用入口（便于GitHub Actions执行）
if __name__ == "__main__":
    import sys
    if len(sys.argv) != 3:
        print("Usage: python stego_core.py <image_path> <project_id>")
        sys.exit(1)
    image_path = sys.argv[1]
    project_id = sys.argv[2]
    try:
        add_watermark(image_path, project_id)
        print(f"Successfully watermarked: {image_path}")
    except Exception as e:
        print(f"Error: {str(e)}")
        sys.exit(1)
