requirements.txt
pillow>=10.0.0
numpy>=1.24.0
flask>=2.3.0
